+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.43.0] - 2019-03-07
 ### Added
 - Add module_options deployment parameter for shared module options
 - Support Terraform backend config via Runway module options
@@ -431,7 +433,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Fix changed CFN parameters not being displayed during `runway plan`.
 
-[Unreleased]: https://github.com/onicagroup/runway/compare/v0.42.0...HEAD
+[Unreleased]: https://github.com/onicagroup/runway/compare/v0.43.0...HEAD
+[0.43.0]: https://github.com/onicagroup/runway/compare/v0.42.0...v0.43.0
 [0.42.0]: https://github.com/onicagroup/runway/compare/v0.41.2...v0.42.0
 [0.41.2]: https://github.com/onicagroup/runway/compare/v0.41.1...v0.41.2
 [0.41.1]: https://github.com/onicagroup/runway/compare/v0.41.0...v0.41.1